pyproject.toml
[project]
name = "backend_endogamia"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AlissonGabrielSantos", email = "alisson.gab.santos@gmail.com"},
]
dependencies = [
    "django>=5.0.2",
    "mysqlclient>=2.2.4",
    "mysql>=0.0.3",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "djangorestframework-simplejwt>=5.3.1",
    "setuptools>=69.1.1",
    "drf-spectacular>=0.27.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
